% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{addFixedDayToCalendar}
\alias{addFixedDayToCalendar}
\title{Title}
\usage{
addFixedDayToCalendar(
  calendar,
  month,
  day,
  weight = 1,
  start = NULL,
  end = NULL
)
}
\arguments{
\item{end}{}
}
\value{

}
\description{
Title
}
