% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{addPrespecifiedHolidayToCalendar}
\alias{addPrespecifiedHolidayToCalendar}
\title{Title}
\usage{
addPrespecifiedHolidayToCalendar(
  calendar,
  event,
  offset = 0,
  weight = 1,
  start = NULL,
  end = NULL
)
}
\arguments{
\item{end}{}
}
\value{

}
\description{
Title
}
