% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{calendar.easter}
\alias{calendar.easter}
\title{Title}
\usage{
calendar.easter(
  calendar,
  offset,
  julian = F,
  weight = 1,
  start = NULL,
  end = NULL
)
}
\arguments{
\item{calendar}{The calendar.}

\item{weight}{}

\item{end}{}
}
\value{

}
\description{
Title
}
