% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{calendar.fixedday}
\alias{calendar.fixedday}
\title{Title}
\usage{
calendar.fixedday(calendar, month, day, weight = 1, start = NULL, end = NULL)
}
\arguments{
\item{end}{}
}
\value{

}
\description{
Title
}
