% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{calendar.fixedweekday}
\alias{calendar.fixedweekday}
\title{Add fixed week day to a calendar}
\usage{
calendar.fixedweekday(
  calendar,
  month,
  week,
  dayofweek,
  weight = 1,
  start = NULL,
  end = NULL
)
}
\arguments{
\item{calendar}{The calendar.}

\item{month}{the month of the day to add: from 1 (January) to 12 (December).}

\item{week}{the number of the week of the day to add: from 1 (first week of the month) to 5.}

\item{dayofweek}{the day of the week: from 1 (Monday) to 7 (Sunday).}

\item{weight}{weight associated to the holiday.}

\item{start, end}{Validity period of the holiday in the format \code{"YYYY-MM-DD"}.}
}
\description{
Add fixed week day to a calendar
}
