% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{calendar.holiday}
\alias{calendar.holiday}
\title{Title}
\usage{
calendar.holiday(
  calendar,
  event,
  offset = 0,
  weight = 1,
  start = NULL,
  end = NULL
)
}
\arguments{
\item{calendar}{The calendar.}

\item{event}{}

\item{weight}{}

\item{end}{}
}
\value{

}
\description{
Title
}
