% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{calendar.singledate}
\alias{calendar.singledate}
\title{Add Single Date to a Calendar}
\usage{
calendar.singledate(calendar, date, weight = 1)
}
\arguments{
\item{calendar}{The calendar.}

\item{date}{the date of the holiday in the format \code{"YYYY-MM-DD"}.}

\item{weight}{weight associated to the holiday.}
}
\description{
Add Single Date to a Calendar
}
