% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{fixed_day}
\alias{fixed_day}
\title{Title}
\usage{
fixed_day(month, day, weight = 1, validity = NULL)
}
\arguments{
\item{validity}{}
}
\description{
Title
}
\examples{
day<-fixed_day(7, 21, .9) # 21 July, with weight=0.9
day
}
