% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{fixed_week_day}
\alias{fixed_week_day}
\title{Title}
\usage{
fixed_week_day(month, week, dayofweek, weight = 1, validity = NULL)
}
\arguments{
\item{month}{Month of the holiday}

\item{week}{Position of the specified week day in the month (3 = third <dayofweek> of the month... ). Should be always lower than 5.
-1 for the last <dayofweek> of the month}

\item{dayofweek}{Day of week 1 for Monday, 7 for Sunday}

\item{validity}{}
}
\description{
Title
}
\examples{
day<-fixed_week_day(7, 2, 3) # second Tuesday of July
day

}
