% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{holidays}
\alias{holidays}
\title{Title}
\usage{
holidays(
  calendar,
  start,
  length,
  nonworking = c(6, 7),
  type = c("Skip", "All", "NextWorkingDay", "PreviousWorkingDay")
)
}
\arguments{
\item{calendar}{The calendar}

\item{start}{First day of the calendar}

\item{length}{Length of the calendar}

\item{nonworking}{Indexes of non working days (Monday=1, Sunday=7)}

\item{type}{Adjustment type when a holiday falls a week-end: "NextWorkingDay",
"PreviousWorkingDay",
"Skip" (holidays corresponding to non working days are simply skipped in the matrix),
"All" (holidays are always put in the matrix, even if they correspond to a non working day)}
}
\value{

}
\description{
Title
}
\examples{
belgiumCalendar<-calendar.new()
calendar.fixedday(belgiumCalendar, 7, 21)
calendar.holiday(belgiumCalendar, "NEWYEAR")
calendar.holiday(belgiumCalendar, "CHRISTMAS")
calendar.holiday(belgiumCalendar, "CHRISTMAS", offset=1, weight=.5)
calendar.holiday(belgiumCalendar, "MAYDAY")
calendar.holiday(belgiumCalendar, "EASTERMONDAY")
calendar.holiday(belgiumCalendar, "WHITMONDAY")
calendar.holiday(belgiumCalendar, "ASSUMPTION")
calendar.holiday(belgiumCalendar, "ALLSAINTDAY")
calendar.holiday(belgiumCalendar, "ARMISTICE")
q<-holidays(belgium, "2021-01-01", 365.25*10, type="NextWorkingDay")
plot(apply(q,1, max))
}
