% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{intervention.variable}
\alias{intervention.variable}
\title{Title}
\usage{
intervention.variable(
  frequency,
  start,
  length,
  starts,
  ends,
  delta = 0,
  seasonaldelta = 0
)
}
\arguments{
\item{seasonaldelta}{}
}
\value{

}
\description{
Title
}
