% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobuf.R, R/arima.R, R/calendars.R,
%   R/modellingcontext.R, R/regarima_rslts.R
\name{jd3_utilities}
\alias{jd3_utilities}
\alias{p2r_arima}
\alias{p2r_ucarima}
\alias{p2r_spec_sarima}
\alias{r2p_spec_sarima}
\alias{p2r_outliers}
\alias{r2p_outliers}
\alias{p2r_ramps}
\alias{r2p_ramps}
\alias{p2r_uservars}
\alias{r2p_uservars}
\alias{p2r_variables}
\alias{r2jd_sarima}
\alias{jd2r_ucarima}
\alias{.r2p_validityPeriod}
\alias{.p2r_validityPeriod}
\alias{.p2r_fixedday}
\alias{.r2p_fixedday}
\alias{.p2r_fixedweekday}
\alias{.r2p_fixedweekday}
\alias{.p2r_easterday}
\alias{.r2p_easterday}
\alias{.p2r_singleday}
\alias{.r2p_singleday}
\alias{.p2r_specialday}
\alias{.r2p_specialday}
\alias{.r2p_holiday}
\alias{.p2r_calendar}
\alias{.r2p_calendar}
\alias{.p2r_chainedcalendar}
\alias{.r2p_chainedcalendar}
\alias{.p2r_wcalendar}
\alias{.r2p_wcalendar}
\alias{.p2r_calendardef}
\alias{.r2p_calendardef}
\alias{.r2p_moniker}
\alias{.p2r_moniker}
\alias{.r2p_datasupplier}
\alias{.p2r_datasupplier}
\alias{.r2p_datasuppliers}
\alias{.p2r_datasuppliers}
\alias{.p2r_modellingcontext}
\alias{.r2p_modellingcontext}
\alias{p2r_regarima_rslts}
\title{Java Utility Functions}
\usage{
p2r_arima(p)

p2r_ucarima(p)

p2r_spec_sarima(spec)

r2p_spec_sarima(r)

p2r_outliers(p)

r2p_outliers(r)

p2r_ramps(p)

r2p_ramps(r)

p2r_uservars(p)

r2p_uservars(r)

p2r_variables(p)

r2jd_sarima(model)

jd2r_ucarima(jucm)

.r2p_validityPeriod(start, end)

.p2r_validityPeriod(vp)

.p2r_fixedday(p)

.r2p_fixedday(r)

.p2r_fixedweekday(p)

.r2p_fixedweekday(r)

.p2r_easterday(p)

.r2p_easterday(r)

.p2r_singleday(p)

.r2p_singleday(r)

.p2r_specialday(p)

.r2p_specialday(r)

.r2p_holiday(r)

.p2r_calendar(p)

.r2p_calendar(r)

.p2r_chainedcalendar(p)

.r2p_chainedcalendar(r)

.p2r_wcalendar(p)

.r2p_wcalendar(r)

.p2r_calendardef(p)

.r2p_calendardef(r)

.r2p_moniker(r)

.p2r_moniker(p)

.r2p_datasupplier(name, r)

.p2r_datasupplier(p)

.r2p_datasuppliers(r)

.p2r_datasuppliers(p)

.p2r_modellingcontext(p)

.r2p_modellingcontext(r)

p2r_regarima_rslts(p)
}
\arguments{
\item{p, r, spec, model, jucm, vp, start, end, jrslts}{parameters.}
}
\description{
These functions are used in all JDemetra+ 3.0 packages to easily interact between R and Java objects.
}
