% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{julianeaster.variable}
\alias{julianeaster.variable}
\title{Title}
\usage{
julianeaster.variable(frequency, start, length, duration = 6)
}
\arguments{
\item{duration}{}
}
\value{

}
\description{
Title
}
