% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{longTermMean}
\alias{longTermMean}
\title{Long-term means of a calendar}
\usage{
longTermMean(calendar, frequency, groups = c(1, 2, 3, 4, 5, 6, 0), holiday = 7)
}
\arguments{
\item{calendar}{The calendar.}

\item{groups}{Groups of days. The length of the array must be 7. It indicates to what group each week day
belongs. The first item corresponds to Mondays and the last one to Sundays. The group used for contrasts (usually Sundays) is identified by 0.
The other groups are identified by 1, 2,... n (<= 6). For instance, usual trading days are defined by c(1,2,3,4,5,6,0),
week days by c(1,1,1,1,1,0,0), week days, Saturdays, Sundays by c(1,1,1,1,1,2,0) etc...}

\item{holiday}{Day for holidays (holidays are considered as that day). 1 for Monday... 7 for Sunday. Doesn't necessary belong to the 0-group}
}
\value{
The long term means corresponding to each group/period, starting with the 0-group
}
\description{
Long-term means of a calendar
}
