% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modellingcontext.R
\name{modelling_context}
\alias{modelling_context}
\title{Title}
\usage{
modelling_context(calendars = NULL, variables = NULL)
}
\arguments{
\item{calendars}{Calendars}

\item{variables}{Variables}
}
\description{
Title
}
\examples{
BE<-national_calendar(list(
    fixed_day(7,21),
    special_day('NEWYEAR'),
    special_day('CHRISTMAS'),
    special_day('MAYDAY'),
    special_day('EASTERMONDAY'),
    special_day('ASCENSION'),
    special_day('WHITMONDAY'),
    special_day('ASSUMPTION'),
    special_day('ALLSAINTSDAY'),
    special_day('ARMISTICE')))
FR<-national_calendar(list(
    fixed_day(5,8),
    fixed_day(7,14),
    special_day('NEWYEAR'),
    special_day('CHRISTMAS'),
    special_day('MAYDAY'),
    special_day('EASTERMONDAY'),
    special_day('ASCENSION'),
    special_day('WHITMONDAY'),
    special_day('ASSUMPTION'),
    special_day('ALLSAINTSDAY'),
    special_day('ARMISTICE')))
#simple list of ts
vars<-list(v1=rjd3toolkit::ABS$X0.2.09.10.M, v2=rjd3toolkit::ABS$X0.2.05.10.M)
MC<-modelling_context(calendars=list(BE=BE, FR=FR), variables<-vars)
}
