% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{periodic.dummies}
\alias{periodic.dummies}
\alias{periodic.contrasts}
\title{Periodic dummies and contrasts}
\usage{
periodic.dummies(frequency, start, length)

periodic.contrasts(frequency, start, length)
}
\arguments{
\item{frequency}{Annual frequency. Should be a divisor of 12.}

\item{start}{Array with the first year and the first period (for instance \code{c(1980, 1)}).}

\item{length}{Length of the variables}
}
\description{
Periodic dummies and contrasts
}
