% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differencing.R
\name{rangemean.tstat}
\alias{rangemean.tstat}
\title{Title}
\usage{
rangemean.tstat(data, period = 0, groupsize = 0, trim = 0)
}
\arguments{
\item{data}{data to test.}

\item{period}{periodicity of the data.}

\item{trim}{}
}
\value{
T-Stat of the range-mean regression.
}
\description{
Function to perform a trimmed range mean regression test used in TRAMO
to select whether the original series will be transformed into log or maintain the level.
}
