% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{stock.td}
\alias{stock.td}
\title{Stock Trading days}
\usage{
stock.td(frequency, start, length, w = 31)
}
\arguments{
\item{frequency}{Annual frequency. Should be a divisor of 12.}

\item{start}{Array with the first year and the first period (for instance \code{c(1980, 1)}).}

\item{length}{Length of the variables}

\item{w}{indicates day of the month when inventories and other stock are reported (to denote the last day of the month enter 31).}
}
\description{
Stock Trading days
}
