% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td.R
\name{td.ch}
\alias{td.ch}
\title{Canova-Hansen Trading Days test}
\usage{
td.ch(s, differencing)
}
\arguments{
\item{s}{a \code{ts} object that corresponds to the input time series to test.}

\item{differencing}{differencing lags}
}
\description{
Canova-Hansen Trading Days test
}
