% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{td.forTs}
\alias{td.forTs}
\title{Usual trading days variables for a given time series}
\usage{
td.forTs(s, groups = c(1, 2, 3, 4, 5, 6, 0), contrasts = T)
}
\arguments{
\item{s}{The time series}

\item{groups}{Groups of days. The length of the array must be 7. It indicates to what group each week day
belongs. The first item corresponds to Mondays and the last one to Sundays. The group used for contrasts (usually Sundays) is identified by 0.
The other groups are identified by 1, 2, ... n (<= 6). For instance, usual trading days are defined by c(1,2,3,4,5,6,0),
week days by c(1,1,1,1,1,0,0), week days, Saturdays, Sundays by c(1,1,1,1,1,2,0) etc...}

\item{contrasts}{If true, the variables are defined by contrasts with the 0-group. Otherwise, raw number of days are provided}
}
\value{

}
\description{
Usual trading days variables for a given time series
}
