% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{trigonometric.variables}
\alias{trigonometric.variables}
\title{Trigonometric variables}
\usage{
trigonometric.variables(frequency, start, length, seasonal_frequency = NULL)
}
\arguments{
\item{frequency}{Annual frequency. Should be a divisor of 12.}

\item{start}{Array with the first year and the first period (for instance \code{c(1980, 1)}).}

\item{length}{Length of the variables}

\item{seasonal_frequency}{the seasonal frequencies.
By default the fundamental seasonal frequency and all the harmonics are used.}
}
\description{
Computes trigonometric variables at different frequencies
}
\details{
\loadmathjax
Denote by \mjseqn{P} the value of \code{frequency} (= the period) and
\mjseqn{f_1}, ..., \mjseqn{f_n} the frequencies provides by \code{seasonal_frequency}
(if \code{seasonal_frequency = NULL} then \mjseqn{n=\lfloor P/2\rfloor} and \mjseqn{f_i}=i).

\code{trigonometric.variables} returns a matrix of size \mjseqn{length\times(2n)}.

For each date \mjseqn{t} associated to the period \mjseqn{m} (\mjseqn{m \in [1,P]}),
the columns \mjseqn{2i} and \mjseqn{2i-1} are equal to:
\mjsdeqn{
\cos \left(
\frac{2 \pi}{P}  \times m \times f_i
\right)
\text{ and }
\sin \left(
\frac{2 \pi}{P} \times m \times f_i
\right)
}
Take for example the case when the first date (\code{date}) is a January, \code{frequency = 12}
(monthly time series), \code{length = 12} and \code{seasonal_frequency = NULL}.
The first frequency, \mjseqn{\lambda_1 = 2\pi /12} represent the fundamental seasonal frequency and the
other frequencies (\mjseqn{\lambda_2 = 2\pi /12 \times 2}, ..., \mjseqn{\lambda_6 = 2\pi /12 \times 6})
are the five harmonics. The output matrix will be equal to:
\mjsdeqn{
\begin{pmatrix}
\cos(\lambda_1) & \sin (\lambda_1) & \cdots &
\cos(\lambda_6) & \sin (\lambda_6) \newline
\cos(\lambda_1\times 2) & \sin (\lambda_1\times 2) & \cdots &
\cos(\lambda_6\times 2) & \sin (\lambda_6\times 2)\newline
\vdots & \vdots & \cdots & \vdots & \vdots \newline
\cos(\lambda_1\times 12) & \sin (\lambda_1\times 12) & \cdots &
\cos(\lambda_6\times 12) & \sin (\lambda_6\times 12)
\end{pmatrix}
}
}
