% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_spec.R
\name{utility-spec}
\alias{utility-spec}
\alias{createVariable}
\alias{createRamp}
\alias{createOutlier}
\title{Utility Functions For Specifications}
\usage{
createVariable(
  id,
  name = NULL,
  lag0 = 0,
  lag1 = 0,
  coef = NULL,
  regeffect = c("Undefined", "Trend", "Seasonal", "Irregular", "Series",
    "SeasonallyAdjusted")
)

createRamp(start, end, name = NULL, coef = NULL)

createOutlier(code, pos, name = NULL, coef = NULL)
}
\arguments{
\item{start, end, name, coef, code, pos, id, lag0, lag1, coef, regeffect}{}
}
\description{
Utility functions to set specifications used in other packages
}
