% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{easter.variable}
\alias{easter.variable}
\alias{julianeaster.variable}
\title{Easter regressors}
\usage{
easter.variable(
  frequency,
  start,
  length,
  s,
  duration = 6,
  endpos = -1,
  correction = c("Simple", "PreComputed", "Theoretical", "None")
)

julianeaster.variable(frequency, start, length, s, duration = 6)
}
\arguments{
\item{frequency}{Annual frequency (divisor of 12).}

\item{start, length}{First date (array with the first year and the first period)
(for instance \code{c(1980, 1)}) and number of periods of the output variables. Can also be provided with the \code{s} argument}

\item{s}{time series used to get the dates for the trading days variables. If supplied the
parameters \code{frequency}, \code{start} and \code{length} are ignored.}

\item{duration}{Duration (length in days) of the Easter effect.}

\item{endpos}{Position of the end of the Easter effect, relatively to Easter.}

\item{correction}{mean correction option.}
}
\description{
Easter regressors
}
