% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{easter.variable}
\alias{easter.variable}
\alias{easter.variable.forTs}
\title{Easter regressors}
\usage{
easter.variable(
  frequency,
  start,
  length,
  duration,
  endpos = -1,
  correction = c("Simple", "PreComputed", "Theoretical", "None")
)

easter.variable.forTs(
  s,
  duration,
  endpos = -1,
  correction = c("Simple", "PreComputed", "Theoretical", "None")
)
}
\arguments{
\item{frequency}{Annual frequency. Should be a divisor of 12}

\item{start}{Array with the first year and the first period (for instance \code{c(1980, 1)}).}

\item{length}{Length of the variables}

\item{duration}{Duration (length in days) of the Easter effect.}

\item{endpos}{Position of the end of the Easter effect, relatively to Easter.}

\item{correction}{}

\item{s}{time series used to get the dates for the trading days variables.}
}
\description{
Easter regressors
}
