% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{easter.variable}
\alias{easter.variable}
\title{Title}
\usage{
easter.variable(
  frequency,
  start,
  length,
  duration,
  endpos = -1,
  correction = c("Simple", "PreComputed", "Theoretical", "None")
)
}
\arguments{
\item{correction}{}
}
\value{

}
\description{
Title
}
