% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{holidays}
\alias{holidays}
\title{Gets the days corresponding to the holidays}
\usage{
holidays(
  calendar,
  start,
  length,
  nonworking = c(6, 7),
  type = c("Skip", "All", "NextWorkingDay", "PreviousWorkingDay"),
  single = FALSE
)
}
\arguments{
\item{calendar}{The calendar}

\item{start}{First day of the calendar in the format \code{"YYYY-MM-DD"}.}

\item{length}{Length of the calendar.}

\item{nonworking}{Indexes of non working days (Monday=1, Sunday=7).}

\item{type}{Adjustment type when a holiday falls a week-end: \code{"NextWorkingDay"},
\code{"PreviousWorkingDay"},
\code{"Skip"} (holidays corresponding to non working days are simply skipped in the matrix),
\code{"All"} (holidays are always put in the matrix, even if they correspond to a non working day).}

\item{single}{Single variable or matrix containing the different holidays in separate columns}
}
\value{
A matrix where each column is associated to a holiday (in the order of creation of the holiday) and each row to a date.
}
\description{
Gets the days corresponding to the holidays
}
\examples{
belgiumCalendar<-calendar.new()
calendar.fixedday(belgiumCalendar, 7, 21)
calendar.holiday(belgiumCalendar, "NEWYEAR")
calendar.holiday(belgiumCalendar, "CHRISTMAS")
calendar.holiday(belgiumCalendar, "CHRISTMAS", offset=1, weight=.5)
calendar.holiday(belgiumCalendar, "MAYDAY")
calendar.holiday(belgiumCalendar, "EASTERMONDAY")
calendar.holiday(belgiumCalendar, "WHITMONDAY")
calendar.holiday(belgiumCalendar, "ASSUMPTION")
calendar.holiday(belgiumCalendar, "ALLSAINTSDAY")
calendar.holiday(belgiumCalendar, "ARMISTICE")
q<-holidays(belgiumCalendar, "2021-01-01", 365.25*10, type="NextWorkingDay")
plot(apply(q,1, max))
}
