% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{htd}
\alias{htd}
\alias{htd.forTs}
\title{Calendar specific trading days variables}
\usage{
htd(
  calendar,
  frequency,
  start,
  length,
  groups = c(1, 2, 3, 4, 5, 6, 0),
  holiday = 7,
  contrasts = TRUE,
  meanCorrection = contrasts
)

htd.forTs(
  s,
  calendar,
  groups = c(1, 2, 3, 4, 5, 6, 0),
  holiday = 7,
  contrasts = TRUE
)
}
\arguments{
\item{calendar}{The calendar.}

\item{frequency}{Annual frequency. Should be a divisor of 12.}

\item{start}{Array with the first year and the first period (for instance \code{c(1980, 1)}).}

\item{length}{Length of the variables}

\item{groups}{Groups of days. The length of the array must be 7. It indicates to what group each week day
belongs. The first item corresponds to Mondays and the last one to Sundays. The group used for contrasts (usually Sundays) is identified by 0.
The other groups are identified by 1, 2,... n (<= 6). For instance, usual trading days are defined by c(1,2,3,4,5,6,0),
week days by c(1,1,1,1,1,0,0), week days, Saturdays, Sundays by c(1,1,1,1,1,2,0) etc...}

\item{holiday}{Day for holidays (holidays are considered as that day). 1 for Monday... 7 for Sunday. Doesn't necessary belong to the 0-group}

\item{contrasts}{If true, the variables are defined by contrasts with the 0-group. Otherwise, raw number of days are provided}

\item{meanCorrection}{boolean indicating if the regressors are corrected for long-term term.
By default the correction is done if \code{contrasts = TRUE}.}

\item{s}{The time series.}
}
\value{
The variables corresponding to each group, starting with the 0-group (\code{contrasts = FALSE})
or the 1-group (\code{contrasts = TRUE}).
}
\description{
Calendar specific trading days variables
}
