% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{intervention.variable}
\alias{intervention.variable}
\title{Intervention variable}
\usage{
intervention.variable(
  frequency,
  start,
  length,
  starts,
  ends,
  delta = 0,
  seasonaldelta = 0
)
}
\arguments{
\item{frequency}{Annual frequency. Should be a divisor of 12.}

\item{start}{Array with the first year and the first period (for instance \code{c(1980, 1)}).}

\item{length}{Length of the variables}

\item{seasonaldelta}{}
}
\value{

}
\description{
Intervention variable
}
