% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{intervention.variable}
\alias{intervention.variable}
\title{Intervention variable}
\usage{
intervention.variable(
  frequency,
  start,
  length,
  s,
  starts,
  ends,
  delta = 0,
  seasonaldelta = 0
)
}
\arguments{
\item{frequency, start, length}{Annual frequency (divisor of 12), first date (array with the first year and the first period)
(for instance \code{c(1980, 1)}) and number of periods of the output variables. Can also be provided with the \code{s} argument}

\item{s}{time series used to get the dates for the trading days variables. If supplied the
parameters \code{frequency}, \code{start} and \code{length} are ignored.}

\item{starts, ends}{characters specifying sequences of starts/ends dates for the intervention variable.
Can be characters or integers.}

\item{delta}{regular differencing order.}

\item{seasonaldelta}{segular differencing order.}
}
\description{
Intervention variable
}
\details{
Intervention variables are combinations of any possible sequence of ones and zeros
(the sequence of ones being defined by  by the parameters \code{starts} and \code{ends})
and of \eqn{\frac{1}{(1-B)^d}} and \eqn{\frac{1}{(1-B^s)^D}} where \eqn{B} is the
backwar operation, \eqn{s} is the frequency of the time series,
\eqn{d} and \eqn{D} are the parameters \code{delta} and \code{seasonaldelta}.

For example, with \code{delta = 0} and \code{seasonaldelta = 0} we get temporary level shifts defined
by the parameters \code{starts} and \code{ends}. With \code{delta = 1} and \code{seasonaldelta = 0} we get
the cumulative sum of temporary level shifts.
}
\examples{
intervention.variable(12, c(2000, 1), 60,
    starts = "2001-01-01", ends = "2001-12-01")
intervention.variable(12, c(2000, 1), 60,
    starts = "2001-01-01", ends = "2001-12-01", delta = 1)
}
