% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobuf.R, R/arima.R, R/calendars.R,
%   R/regarima_rslts.R
\name{jd3_utilities}
\alias{jd3_utilities}
\alias{p2r_arima}
\alias{p2r_ucarima}
\alias{p2r_spec_sarima}
\alias{r2p_spec_sarima}
\alias{p2r_outliers}
\alias{r2p_outliers}
\alias{p2r_ramps}
\alias{r2p_ramps}
\alias{p2r_variables}
\alias{r2jd_sarima}
\alias{jd2r_ucarima}
\alias{r2p_validityPeriod}
\alias{p2r_validityPeriod}
\alias{p2r_fixedday}
\alias{r2p_fixedday}
\alias{p2r_regarima_rslts}
\title{Java Utility Functions}
\usage{
p2r_arima(p)

p2r_ucarima(p)

p2r_spec_sarima(spec)

r2p_spec_sarima(r)

p2r_outliers(p)

r2p_outliers(r)

p2r_ramps(p)

r2p_ramps(r)

p2r_variables(p)

r2jd_sarima(model)

jd2r_ucarima(jucm)

r2p_validityPeriod(start, end)

p2r_validityPeriod(vp)

p2r_fixedday(p)

r2p_fixedday(r)

p2r_regarima_rslts(p)
}
\arguments{
\item{p, r, spec, model, jucm, vp, start, end}{parameters.}
}
\description{
These functions are used in all JDemetra+ 3.0 packages to easily interact between R and Java objects.
}
