% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{longTermMean}
\alias{longTermMean}
\title{Long-term means of a calendar}
\usage{
longTermMean(calendar, frequency, groups = c(1, 2, 3, 4, 5, 6, 0))
}
\arguments{
\item{groups}{}
}
\value{
The long term means corresponding to each group/period, starting with the 0-group
}
\description{
Long-term means of a calendar
}
