% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modellingcontext.R
\name{modelling_context}
\alias{modelling_context}
\title{Create context}
\usage{
modelling_context(calendars = NULL, variables = NULL)
}
\arguments{
\item{calendars}{list of calendars.}

\item{variables}{list of variables.}
}
\value{
list of calendars and variables
}
\description{
Create context
}
\examples{
ts <- ABS$X0.2.09.10.M
xvar <- td(12, start=start(ts),length = length(ts),groups = c(1,1,1,1,1,0,0))
context <- modelling_context(calendars = NULL, variables = list(xvar = xvar))

}
