% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{periodic.dummies}
\alias{periodic.dummies}
\alias{periodic.contrasts}
\title{Periodic dummies and contrasts}
\usage{
periodic.dummies(frequency, start, length, s)

periodic.contrasts(frequency, start, length, s)
}
\arguments{
\item{frequency}{Annual frequency (divisor of 12).}

\item{start, length}{First date (array with the first year and the first period)
(for instance \code{c(1980, 1)}) and number of periods of the output variables. Can also be provided with the \code{s} argument}

\item{s}{time series used to get the dates for the trading days variables. If supplied the
parameters \code{frequency}, \code{start} and \code{length} are ignored.}
}
\description{
Periodic dummies and contrasts
}
