% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{sarima.model}
\alias{sarima.model}
\title{Title}
\usage{
sarima.model(
  period,
  phi = NULL,
  d = 0,
  theta = NULL,
  bphi = NULL,
  bd = 0,
  btheta = NULL,
  name = NULL
)
}
\arguments{
\item{name}{}
}
\value{

}
\description{
Title
}
